rename InstallationDoc - MdxComponents
