+++ src/docs/components/MdxComponents/DocPre.tsx
@@ -4,9 +4,7 @@ import * as React from "react";
 import { CopyButton } from "@components/CopyButton";
 import { cn } from "@utils/cn";
 
-export const ProsePre: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => {
+export const DocPre = ({ children }: { children: React.ReactNode }) => {
   const [code, setCode] = React.useState("");
   return (
     <div className="flex items-center justify-between overflow-x-auto rounded-xl border border-solid border-dark-700 bg-[#011627] px-5 py-3 text-sm">
