+++ src/docs/content/components/badges.mdx
@@ -5,7 +5,7 @@ Badge like component
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="badges" />
