+++ src/docs/content/components/buttons.mdx
@@ -5,7 +5,7 @@ Drive user action with clickable buttons
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="buttons" />
