+++ src/docs/content/components/dividers.mdx
@@ -5,7 +5,7 @@ Create visual separation for a clean, professional look
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="dividers" />
