+++ src/docs/content/components/emojis.mdx
@@ -5,7 +5,7 @@ Add personality and emotion to your messages
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="emojis" />
