+++ src/docs/content/components/hero-sections.mdx
@@ -5,7 +5,7 @@ Beautiful headers to showcase your message
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="hero-sections" />
