+++ src/docs/content/components/lists.mdx
@@ -5,7 +5,7 @@ All lists components
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="lists" />
