+++ src/docs/content/components/social-icons.mdx
@@ -5,7 +5,7 @@ Increase engagement with clickable social icons
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="social-icons" />
