+++ src/docs/content/components/texts.mdx
@@ -5,7 +5,7 @@ Craft compelling content with easy-to-read paragraphs
 import {
   ComponentDemo,
   ComponentSourceTabs,
-} from "src/docs/components/InstallationDocs";
+} from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="texts" />
